# PERF: use copy=False in constructor in unstack (CoW)