@rhshadrach 2024-02-16
> I _think_ we are sure that the `values` from `self.get_new_values(..)... are always new values owned by us`

I don't think that's always the case. E.g. one of the paths is:

```
        if values.ndim == 1:
            values = values[:, np.newaxis]
```

This will be the original objects memory.

```
arr = np.array([1, 2, 3])
newarr = arr[:, np.newaxis]
print(np.shares_memory(arr, newarr))
# True
```